# openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Task API",
    "version": "1.0.0",
    "description": "A simple in-memory Task CRUD API built with Express"
  },
  "servers": [
    { "url": "http://localhost:3000" }
  ],
  "paths": {
    "/": {
      "get": {
        "summary": "API root",
        "responses": {
          "200": { "description": "API info" }
        }
      }
    },
    "/health": {
      "get": { "summary": "Health check", "responses": { "200": { "description": "OK" } } }
    },
    "/tasks": {
      "get": {
        "summary": "List tasks",
        "parameters": [
          { "name": "done", "in": "query", "schema": { "type": "boolean" } },
          { "name": "search", "in": "query", "schema": { "type": "string" } }
        ],
        "responses": { "200": { "description": "A list of tasks" } }
      },
      "post": {
        "summary": "Create a task",
        "requestBody": {
          "required": true,
          "content": { "application/json": { "schema": { "type": "object", "properties": { "title": { "type": "string" } }, "required": ["title"] } } }
        },
        "responses": { "201": { "description": "Created" }, "400": { "description": "Bad Request" } }
      }
    },
    "/tasks/{id}": {
      "get": {
        "summary": "Get a task",
        "parameters": [{ "name": "id", "in": "path", "required": true, "schema": { "type": "integer" } }],
        "responses": { "200": { "description": "Task" }, "404": { "description": "Not found" } }
      },
      "put": {
        "summary": "Update a task",
        "parameters": [{ "name": "id", "in": "path", "required": true, "schema": { "type": "integer" } }],
        "requestBody": { "required": true, "content": { "application/json": { "schema": { "type": "object", "properties": { "title": { "type": "string" }, "done": { "type": "boolean" } } } } } },
        "responses": { "200": { "description": "Updated" }, "400": { "description": "Bad Request" }, "404": { "description": "Not found" } }
      },
      "delete": {
        "summary": "Delete a task",
        "parameters": [{ "name": "id", "in": "path", "required": true, "schema": { "type": "integer" } }],
        "responses": { "204": { "description": "No Content" }, "404": { "description": "Not found" } }
      }
    },
    "/reset": {
      "post": { "summary": "Reset tasks to initial seed", "responses": { "200": { "description": "Reset" } } }
    }
  }
}